fix(rome_js_syntax): Don't panic on unclosed block comment (#2933)
* fix(rome_js_syntax): Don't panic on unclosed block comment

This fixes an issue where Rome panicked when parsing suppression comments from a block comment without the closing `*/`.

This PR gracefully handles the case where a block comment has no closing `*/` by trying to remove a trailing `*` or `/` at the end of the comment and otherwise take the comment as is.

## Tests

I verified that running rome over rome classic no longer panics.

This is an analyzer specific issue. I'm not sure how to best write a test for this issue without adding it to a specific analyzer test. Please advice ;)

* add unit test