feat(rome_js_formatter): Group return type with parameters
Implement the same heuristic as Prettier for grouping parameters with the return type annotation.

This ensures that a return type will break FIRST before the parameters and the parameters only break if they don't fit on the same line together with the expanded return type annotation.

This PR further unifies the formatting of methods and functions to achieve better code reuse.