feat(rome_formatter): `Dedent` IR
This PR adds a new IR that allows to reduce the indention for some content nested inside of an `align` or `indent`.

The behaviour of the element depends on whether the closest ancestor is an `align` or `indent`

* `align`: Removes the `align` spaces
* `indent`: Reduces the indention level by one *EXCEPT* if the level is 0, in which case this is a no-op

This is necessary for the formatting of nested conditional expressions where the `consequent` gets wrapped in an `align` but any nested consequent should be indented by one level (and not an indent). That's why the formatter uses `deden(indent(...))` to undo the `align` but enforce an `indent`.