feat(rome_formatter): `align > indent` nesting with `IndentStyle::Space`
This changes the behaviour of the printer if an `indent` is nested inside an `align`.
The existing behaviour was to convert the `align` to an `indent`.

This PR changes the behaviour to keep the spaces as is, so that an `align(2) > indent` results in a `______` sequence with `IndentStyle::Space(4)`

This PR further improves `Memoized` so that it avoids creating an `Interned` element on the heap if it has no content.