feat(rome_js_syntax): Node factories for unknown nodes
This PR adds node factory methods for unknown nodes. This is certainly more of an edge case, but I have a use case in the formatter when encountering an invalid parenthesized expression in which case the node should be replaced
as an unknown node so, for as far as the formatter is concerned, parenthesized expressions do not exist.

The new API can be used like this:

```rust
let unknown = rome_js_factory::make::js_unknown_expression(
    root.slots().map(|slot| slot.into_syntax_element()),
);
 ```

## Alternatives

An alternative would be to directly use the `SyntaxNode::new_detached` APIs. This works but is less accessible and
it might be convenient to search for all places where we construct unknown nodes.