feat(rome_formatter): Dedent Root IR
This PR extends the `dedent` IR element to support dedenting by one level or to the start of the line.

I intentionally called the mode `root` because it may become necessary in the future that it's possible to mark a root which is different to the start of the line.

This IR is necessary to correctly format `TemplateLiteral`s where inserting any indent characters that weren't present of the template in the source changes the `template`s content at runtime.