refactor(rome_formatter): Extract `FormatOptions`
## Summary

This PR extracts a `FormatOptions` trait from the `FormatContext`.

The reason for splitting `FormatOptions` and `FormatContext` is the context is intended to store data that is relevant for the formatting of an object like the extracted comments, a source map etc. The information about source maps, comments etc. things that consumers shouldn't be concerned about.

The `FormatOptions` on the other hand are relevant for users because it allows them to customize the formatting of an object.

## Tests

This PR isn't adding any new behaviour.

`cargo test`