refactor(rome_formatter): `FormatLanguage` trait
This PR introduces a new `FormatLanguage` trait that formatters for a specific language must implement.

The `format_*` function family in the `rome_formatter` have rather complicated type parameters that Rust isn't always able to infer and some have a large number of parameters. Making this worse, new features like the pre-processing of AST nodes and the comments refactoring will further increase the parametrization of these functions.

This is why this PR introduces the new `FormatLanguage` trait that groups the language specific types and expose methods to customize the formatter's behaviour.