feat(rome_rowan): `cast_ref` and `match_ast` macro
This PR introduces a new `AstNode::cast_ref` method that takes a node by reference and only clones it if it can be cast into the `AstNode`.

Being able to pass a reference instead of the owned value can be useful if it's necessary to test against multiple nodes

```
if let Some(if_statement) = JsIfStatement::cast(node.clone()) {
  ..
} else if let Some(expression_stmt) = JsExpressionStatement::cast(node.clone()) {
  ..
} else if let Some(switch_statement) = JsSwitchStatement::cast(node) {
  ..
}
```

As you can see here, the `node` must be cloned `n-1` times where `n` is the number of branches (Rust may be able to optimize them away but doesn't hurt to help the compiler). Using `cast_ref` reduces the clones to at most one in the case the node matches one branch.

This PR further introduces a `match_ast` macro that allows matching a syntax node against AST nodes. This removes the need to match on `JsSyntaxKind` directly and use `unwrap_cast` all over the code base.

```rust
    match_ast! {
        match parent {
            // Only allows expression in the `object` child.
            JsStaticMemberExpression(_) => true,
            JsStaticMemberAssignment(_) => true,
            JsComputedMemberExpression(member_expression) => {
                 member_expression
                    .object()
                    .map(ExpressionNode::into_resolved_syntax)
                    .as_ref()
                    == Ok(node)
            },
            JsComputedMemberAssignment(assignment) => {
                assignment
                    .object()
                    .map(ExpressionNode::into_resolved_syntax)
                    .as_ref()
                    == Ok(node)
            },
            _ => false,
        }
    }
```

## Tests

I added a doc test to the new macro.