feat(rome_formatter): `IndentIfBreak` IR element
This PR adds a new `indent_if_break` IR element that has the same semantics as using `if_group_breaks` with `indent` and `if_group_fits_on_line`

```
write!(f, [
  group(&text("Head")).with_group_id(Some(id)),
  if_group_breaks(&indent(&text("indented"))).with_group_id(Some(id)),
  if_group_fits_on_line(&text("indented")).with_group_id(Some(id))
])
```

The advantage of the new IR is that it simplifies the debug output of the IR and removes the need to memoize the content.

## Tests

I added new doc tests and ran `cargo test` to verify that using `indent_if_break` doesn't introduce any regression on existing syntax.