refactor(rome_formatter): Return reference from `context.comments()`
`context.comments` used to return a `Rc<Comments>` so that the lifetime of `comments` isn't bound to the lifetime of the `FormatContext`. This is necessary to support writing to the formatter while holding to a reference of comments:

(Pseudo code):

```rust
let comments = f.context().comments();

for comment in comments.leading_comments(node) {
  write!(f, [comment])?;
}
```

The downside of returning an `Rc` is that every code path using `comments` pays for the overhead of incrementing the `Rc` counter.

This PR changes `Comments` to use an `Rc` internally to enable cheap cloning, similar to how `SyntaxNode` works. This allows `context.comments()` to return a reference to `Comments` instead and only code paths that need to write to the formatter have to pay the overhead of cloning the comments.

## Tests

`cargo test` as this PR isn't adding any new functionality