perf(rome_formatter): Make source maps optional
The Printer source maps are only needed for range formatting but the Printer used to generate them all the time. This has an even higher cost since we rewrite the source tree because the markers must be mapped back to the original position in the source tree.

This PR requires consumers to explicitly enable source maps if they need the source mapping information.