feat(rome_js_formatter): Conditional JSX Chain layout
This PR implements Prettier's conditional JSX chain layout.

```
abcdefgh ? (
	<Element>
		<Sub />
		<Sub />
	</Element>
) : (
	<Element2>
		<Sub />
		<Sub />
	</Element2>
);
```

That parenthesizes the `consequent` and `alternate` except for `null`, `undefined`, and nested conditionals in the alternate.

This PR further removes the `Default` implementation from `JsFormatOptions` because passing a `SourceType` that matches with the type of the AST is mandatory to get correct formatting results.

## Tests

I verified the prettier snapshots and added our own tests to cover the exceptions of `null`, `undefined` and nested alternates that don't get parenthesized.