test(rome_js_formatter): Make `source_type` mandatory on `JsFormatOptions`
This PR removes the `Default` implementation from `JsFormatOptions` because passing a `SourceType` that matches with the type of the AST is mandatory to get correct formatting results.

I noticed that we aren't always passing in the right source type when adding another test to only apply some expensive logic if the source type is JSX and it never took that branch.

## Tests

See now changed prettier tests