feat(rome_formatter): Source Map iteration over deleted ranges
This PR adds a `DeletedRanges` iterator to the `TransformedSourceMap` that allows iterating over all deleted ranges including the deleted text.

The comments refactoring needs a way to identify if a trivia piece is right before the closing `)` in the original tree because we want to keep trailing comments inside parentheses with the preceding nodes

```javascript
!(
  a
  // some comment
);
b
```

The comment should be a trailing comment of `a` and not become a leading comment of `b` because the transformed tree has the following shape

```javascript
!a
  // Some comment
b;
```

Such an iterator requires a fast way to get the text of a deleted range. This isn't possible with `SyntaxNodeText` because resolving the text requires iterating to the token with the given offset which is `O(n)`. That's why this PR changes the `TransfromedSourceMap` to hold a `String` with the original source text to get `O(1)` access to the original text.

The downside of doing so is that the source text must be built up insideof the `SyntaxRewriter` and is stored twice in memory.

The cost of traversal should offset the cost in the comments builder and the memory allocation is short-lived, not adding a lot of pressure if Rome's running as a server.

## Tests
I added a new unit test that shows how the new iterator can be used