perf(rome_formatter): Use `TextSize` in printer
Use `TextSize` instead of usize to represent text positions in the printer.

This PR also fixes an issue where the printer panics if group ids aren't written to the document in the same order as they have been created because `resize` shrinks a vector if its size exceeds the argument.