fix(rome_formatter): Best fitting start entry
I noticed this when investigating how to correctly implement `fill`.

The problem with only pushing the items of the variant after the `StartEntry` is that
any `QueuePredicate` that counts the start/end entries won't see the corresponding
`StartEntry`.

Manually pushing the `StartEntry` onto the stack isn't even necessary because it uses the same `args` anyway