fix(rome_js_parser): Export `{"a"} from`
This PR fixes an issue in our parser where the following syntax was incorrectly rejected.

```js
export {"a"} from "b";
```

The reason why it was rejected is that Rome parses the statement as an `export {...}` statement first where `{"a"}` isn't valid and re-parses the statement when it ends at a `from` statement.

But this means that the parsing of `export {...}` from must be at least as permissive as `export {...} from` to guarantee that the parser reaches the `from` keyword.

This PR ensures that this is the case by explicitly handling the `export {"a"};` case