fix(rome_js_formatter): Fix dropped initializer comments
This PR fixes an issue where comments associated with initializers in variable declarators or class properties were dropped. I noticed this when formatting the `vscode` repo.

## Tests

I added a bunch of new snapshot tests to verify the placement of comments for properties and variable declarators (fewer for variable declarators, because there are already many in Prettier's test suite).

I also run Rome format on the vscode repo to verify that there are no more diagnostics about dropped comments