fix(rome_js_formatter): Arrow chain trailing comments
This PR fixes an issue where trailing comments of arrow expressions that are part of an arrow chain (arrow expressions that are the body of another arrow) were dropped.

The manual handling of comments is necessary because the arrow chain formatting doesn't call into the `FormatJsArrowFunctionExpression` of nested arrow functions.

The implementation already handles the formatting of leading comments correctly.

## Tests

I added a new snapshot test and verified that running rome format on prettier yields no more "dropped comments" errors.