fix(rome_js_formatter): Fix arrow binding parameter in test call
Rome has a special formatting for test calls that ensures that callbacks remain on the same line as the `it` or `describe` even if exceeding the configured print width

```
it('should have the default duration when using the onClose arguments', done => {
	expect(true);
	done();
});
```

This logic has been missing for arrow functions with a single binding parameter (not using parameters). This PR adds the additional check.

## Tests

I added a new snapshot test and verified that formatting the ant-design project twice yields no more formatting changes.