fix(rome_js_parser): call expression in optional member chain followed by `<`
This PR fixes a false diagnostic for call expressions that are part of an optional chain (static member) that are then followed by a binary `<` expression.

```
String(item)?.b <aBcd;
```

The problem is that the error recovery code for `a.test?.` incorrectly kicked in which should not happen if what followed after the call expression is a valid static member expression in an optional chain.

fixes #3486