fix(rome_js_parser): Ignore bindings in desctructuring initializers
Rome's parser implements a very limited check for duplicated bindings by testing if there are no duplicate bindings in the same variable declaration.

This PR fixes an issue where the parser flagged a binding with the same name inside a destructuring initializer as a duplicated binding which is incorrect.

Fixes #3521