doc(rome_diagnostics): Move Writing Diagnostics to the top
and fix a link to the technical philosophy.

The main motivation behind moving the writing to the top is because I want to link from it from the blog post but it being at the bottom means the browser can't scroll to the top of the section.

The second reason is that it's probably good to know what a diagnostic should be before learning all the technical jargon and details.