fix(rome_cli): Respect formatter/linter `enabled` from configuration
Fixes an issue where `rome ci` ignored the `formatter.enabled` and `linter.disabled` configuration. The issue was that the CI defaulted to `true` when the `--formater/linter-enabled` options weren't specified.

## Test Plan

The reason why the existing integration test didn't work is that Rome doesn't check the formatting for files that have lint/parse errors. I changed the integration test to a file that only has incorrect formatting.