Add context about Romes philosophy
As discussed in https://github.com/rome/tools/discussions/3698#discussioncomment-4128011 if you are new to the world of formatting it might be hard to understand at first why Rome (and Prettier) are so opinionated. Prettier has a whole philosophy page which might be overkill at this point. But adding just a few of these links add's a lot of necessary context as they will either immediately resonate or alienate people which are looking for a less opinionated tool.