fix(rome_js_formatter): Unnecessary parentheses for return/unary with verbatim argument
## Summary
This PR fixes #3735 by only adding parentheses around the argument if the argument isn't suppressed. This is necessary because the verbatim formatting of the argument includes the parentheses too. Formatting the parentheses in the return/unary formatting and as part of the verbatim results in duplicating the parentheses on every save.

## Test Plan
I added a handful of new tests covering the problematic cases.