fix(rome_cli): Fix max diagnostic counts
This PR fixes the max diagnostics count for the linter.

The issue was that `process_file` reports a `Success` if a file has no diagnostic even if the file has skipped diagnostics.

This PR changes `process_file` to return a `Message` if a file has reported diagnostics or any skipped diagnostics.