feat(rome_js_analyze): Improve generated suppression text
The suppression code action used to insert a suppression comment with the *reason* `suppressed`. The issue with `suppressed` is that users might think that this is the exact syntax necessary for suppression comments rather than it being a placeholder for the reason why the rule is suppressed.

This PR changes the inserted suppression:

```
//·rome-ignore·lint/a11y/useKeyWithClickEvents:·<explanation>

```

which hopefully makes it clear to users that they should replace `<explanation>` with a reason.

A more involved solution could select the `<explanation>` after adding the suppression comment.

## Test Plan

See updated snapshots