feat(rome_json_formatter): Fallback to Verbatim for nodes with syntax errors
Some prettier tests for JSON5 are failing after implementing array formatting in #4064

The tests fail because the formatting returns a `FormatError::SyntaxError` if a mandatory child node is missing but the error is never handled by the JSON formatting.

This PR fixes this by

* Gracefully handle missing values on the root value by calling into verbatim formatting if that's the case
* Use `format_or_verbatim` when formatting values

I had to pull out the `format_or_verbatim` from the `rome_js_formatter` and move it to `rome_formatter` so that it can be re-used between js and json formatting.