Replace unsafe rome_diagnostics::Error implementation with double-Box
The stated reason for this `unsafe` code was to avoid bloating the
size of `Result<T, Error>`, since `Box<dyn Diagnostic>` would be a fat
pointer.  But `Box<Box<dyn Diagnostic>>`, a thin pointer to a fat
pointer, acomplishes the same goal safely.  We don’t have millions of
`Error`s in the common case, so this seems like a better tradeoff than
maintaining `unsafe` code here.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>